remove auto, examples, and move controlboard
