+++ src/main/include/commands/DrivetrainCommand.h
@@ -4,7 +4,7 @@
 #include <frc2/command/CommandHelper.h>
 
 #include "subsystems/Drivetrain.h"
-#include "subsystems/ControlBoard.h"
+#include "ControlBoard.h"
 
 /**
  * An example command that uses an example subsystem.
@@ -16,12 +16,6 @@
 class DriveCommand
     : public frc2::CommandHelper<frc2::CommandBase, DriveCommand> {
  public:
-  /**
-   * Creates a new ExampleCommand.
-   *
-   * @param subsystem The subsystem used by this command.
-   */
-
   // DriveCommand(Drivetrain* drivetrain, std::function<double()> forward,
   // std::function<double()> rotation);
   DriveCommand(Drivetrain& drivetrain, ControlBoard& controlboard);
